various changes
- Walkthrough: Add the writing "Artist search..." in the search bar
when a user hasn't typed in anything. Just like in the past view.
- Walkthrough: Change the formatting of search results to look exactly
like the search results on past or upcoming views.
- Walkthrough: Change the message on the second screen to "Find a
concert you attended in (x city):"
- Change the venue names to location names for the hardcoded concerts
on the first screen
- Prevent users from tilting their phones and widening the screen
during the walkthrough. (keep on portrait mode)
-Friend tagging: Remove the plus symbol in the friend row. It will only
appear if they haven't tagged anyone with the message "Tag some of your
friends that went"
- Friend tagging: Long names get cut off. There should be room on
two/three lines just like the artist lineup
- Change the message for the friends row when no one's been tagged from
"tag some friends that went" to "Tag some friends that are going"
- Friend tagging: Add a popup right after they select friends and
before the facebook invite window that says: "x of your friends haven't
joined Encore. Invite them to share this concert with them." With two
options, "Invite" and "Cancel".